20250220-drf-exceptions/pyproject.toml
[project]
name = "20250220-drf-exceptions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-environ>=0.12.0",
    "djangorestframework>=3.15.2",
]
